Add jwt user authentication and token authorization

+++ db/repository/blog.py
@@ -26,7 +26,9 @@ def update_blog(id: int, blog: BlogUpdate, author_id: int, db: Session):
     """"""
     blog_in_db = db.query(Blog).filter(Blog.id == id).first()
     if not blog_in_db:
-        return
+        return {"error":f"Blog with id {id} does not exist"}
+    if not blog_in_db.author_id == author_id:
+        return {"error":f"Only the author can modify the blog"}
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content
     db.add(blog_in_db)
@@ -37,7 +39,9 @@ def delete_blog(id: int, author_id: int, db: Session):
     """"""
     blog_in_db = db.query(Blog).filter(Blog.id == id)
     if not blog_in_db.first():
-        {"error": "Could not find blog with given id {id}"}
+        return {"error": "Could not find blog with given id {id}"}
+    if not blog_in_db.first().author_id == author_id:             #new
+        return {"error":f"Only the author can delete a blog"}
     blog_in_db.delete()
     db.commit()
     return {"msg": f"Successfully deleted blog with id {id}"}
